+++ composeApp/src/commonMain/kotlin/com/virtualworld/mipymeanabelmaster/id/DataModule.kt
@@ -1,6 +1,6 @@
 package com.virtualworld.mipymeanabelmaster.id
 
-import com.virtualworld.mipymeanabelmaster.domain.GetOrdersSentUseCase
+import com.virtualworld.mipymeanabelmaster.domain.usecase.GetOrdersSentUseCase
 import org.koin.core.module.dsl.factoryOf
 import org.koin.dsl.module
