+++ composeApp/src/commonMain/kotlin/com/virtualworld/mipymeanabelmaster/screen/main/MainScreen.kt
@@ -5,18 +5,15 @@ import androidx.compose.foundation.layout.WindowInsets
 import androidx.compose.foundation.layout.calculateEndPadding
 import androidx.compose.foundation.layout.calculateStartPadding
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.waterfall
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Surface
-import androidx.compose.material3.Text
 import androidx.compose.material3.TopAppBarDefaults
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.input.nestedscroll.nestedScroll
 import androidx.compose.ui.unit.LayoutDirection
-import androidx.compose.ui.unit.dp
 import androidx.navigation.compose.rememberNavController
 import com.virtualworld.mipymeanabelmaster.navigation.AppNavHost
 
@@ -32,7 +29,7 @@ fun MainScreen() {
 
         Scaffold(
             // bottomBar = { BottomBarView(navController) },
-            topBar = { TopBarView(scrollBehavior) },
+            topBar = { TopBarView(scrollBehavior,navController) },
             modifier = Modifier.nestedScroll(scrollBehavior.nestedScrollConnection),
             contentWindowInsets = WindowInsets.waterfall
         ) { paddingValues ->
